+++ R/cmstatr.R
@@ -1,0 +1,6 @@
+#' @description
+#' To learn more about \code{cmstatr}, start with the vignettes:
+#' \code{browseVignettes(package = "cmstatr")}
+#' @keywords internal
+#' @docType package
+"_PACKAGE"