+++ NEWS.md
@@ -1,3 +1,7 @@
+# Version 0.5.2
+- Internally use `vapply` instead of `sapply` to improve code safety
+- Increased coverage of unit tests
+
 # Version 0.5.1
 - Fixed the title of the graphing vignette